Support Django-1.7 and drop Django-1.6 support
Core accounts are now created in a data migration, how it's meant to be. The maintainability problem
of manually createing treebeard nodes remains unsolved. Hopefully it will not be much of a problem
in the future.

When you change the settings used in the migration after you ran the migration, you have to write a
corresponding data migration that applies the new settings yourself.